node: Move doc stuff to its own module
